- SeriesMatching::reduceDeformations now should also work for non-rectangular domains - removed "deformationModel 11" from matchSeries.cpp, it's unused and not compatible with the change in SeriesMatching::reduceDeformations - added new registration mode 21 (non-rigid NCC+Laplace on a non-dyadic grid) - added new registration mode 22 (non-rigid SSD+Dirichlet)
updated QuocMesh base:
- fixed numerous clang and GCC warnings
- fixed a compile problem on arm
- exported GradientDescentType for some classes
- Fixed: qc::transformAndClipCoord didn't clip properly on non-square domains.
- fixed some data types in qc::ScalarArray<_DataType, qc::QC_2D>::loadTIFF (uint32 -> uint32_t, uint16 -> uint16_t)
- small change in qc::ScalarArray<DataType, qc::QC_2D>::interpolate to prevent NaNs/Infs from "creeping" in one boundary case
- improved an exception message
- Fixed: testSumWillFit failed for uint64_t under ARM.
- qc::ScalarArray<_DataType, qc::QC_2D>::loadTIFF now throws an exception if TIFFOpen returns NULL
- cmake: Simplified USE_OPENMP, which should now also work with more compilers now, but requires a non-ancient cmake version.
- fixed a problem with more recent SuiteSparse versions
- fixed some CIMG related compile problems
- fixed a compile problem with clang 16
- fixed argument order
- recent versions of VC++ support snprintf
- cmake: Fixed: Adding /bigobj under MSVCx64 didn't work properly.
- cmake: Updated the zlib download link and upgraded the version to 1.2.13.
- small adjustment for SuiteSparse 5.2.0
- fixed out-of-bounds access in aol::MultiVectorWithConcatData for components of size 0
- updated the download link for wavelet-03.rar
- cmake: Improved SuiteSparse detection under Linux.

+++ quocmesh/modules/quoc/dm3Import.h
@@ -138,7 +138,7 @@ public:
 
     // get length of file:
     _in.seekg ( 0, _in.end );
-    const int length = _in.tellg();
+    const size_t length = _in.tellg();
     _in.seekg ( 0, _in.beg );
     const string outFileName = ( OutBaseName + ( ( _version == 3 ) ? ".dm3" : ".dm4" ) );
     std::ofstream out ( outFileName.c_str(), ios::binary );
@@ -155,7 +155,7 @@ public:
       writeValue ( out, getDataEntry().dataType, paddedArray[i] );
 
     _in.seekg ( out.tellp(), _in.beg );
-    for ( int i = out.tellp(); i < length; ++i )
+    for ( size_t i = out.tellp(); i < length; ++i )
       out.put ( _in.get() );
   }
   
@@ -165,7 +165,7 @@ public:
     
     // get length of file:
     _in.seekg ( 0, _in.end );
-    const int length = _in.tellg();
+    const size_t length = _in.tellg();
     _in.seekg ( 0, _in.beg );
     const string outFileName = ( OutBaseName + ( ( _version == 3 ) ? ".dm3" : ".dm4" ) );
     std::ofstream out ( outFileName.c_str(), ios::binary );
@@ -182,7 +182,7 @@ public:
       writeValue ( out, getDataEntry().dataType, paddedArray[i] );
     
     _in.seekg ( out.tellp(), _in.beg );
-    for ( int i = out.tellp(); i < length; ++i )
+    for ( size_t i = out.tellp(); i < length; ++i )
       out.put ( _in.get() );
   }
 protected:
